0.2.0 (#23)
* 📌 Pin deps for docs

* Don't link non-existing files for misc.Glob2Dir

* Upgrade datar to 0.8.*

* ⬆️ Upgrade pipen to v0.3

* ⚡️ Load 10X TCR and RNA-seq data files more robustly for scrna.SeuratPreparing and tcr.ImmunarchLoading

* 🔖 0.2.0

* 🐛 Default value to None when config key doesn't exist

+++ biopipen/namespaces/misc.py
@@ -26,7 +26,9 @@ class Glob2Dir(Proc):
     output = "outdir:dir:from_glob"
     script = """
         for infile in {{in.pattern}}; do
-            ln -s $(realpath $infile) "{{out.outdir}}/$(basename $infile)";
+            if [[ -e $infile ]]; then
+                ln -s $(realpath $infile) "{{out.outdir}}/$(basename $infile)";
+            fi
         done
     """
